Move mgradm (.*) podman to mgradm \1
Since mgradm will not managed Kubernetes deployments, the `podman`
subcommand is removed and converted into an alias for backwards
compatibility.
