+++ mgradm/cmd/install/flags.go
@@ -1,8 +1,8 @@
-// SPDX-FileCopyrightText: 2025 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package shared
+package install
 
 import (
 	"github.com/spf13/cobra"
