+++ mgradm/cmd/install/install_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package podman
+package install
 
 import (
 	"os"
