+++ mgradm/cmd/migrate/utils.go
@@ -1,8 +1,8 @@
-// SPDX-FileCopyrightText: 2025 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package podman
+package migrate
 
 import (
 	"errors"
@@ -10,20 +10,26 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/uyuni-project/uyuni-tools/mgradm/shared/podman"
-	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
-	shared_podman "github.com/uyuni-project/uyuni-tools/shared/podman"
+	podman_utils "github.com/uyuni-project/uyuni-tools/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
+
+	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
 )
 
-var systemd shared_podman.Systemd = shared_podman.NewSystemd()
+var systemd podman_utils.Systemd = podman_utils.NewSystemd()
 
-func upgradePodman(_ *types.GlobalFlags, flags *podmanUpgradeFlags, cmd *cobra.Command, _ []string) error {
-	hostData, err := shared_podman.InspectHost()
+func migrateToPodman(
+	_ *types.GlobalFlags,
+	flags *podmanMigrateFlags,
+	cmd *cobra.Command,
+	args []string,
+) error {
+	hostData, err := podman_utils.InspectHost()
 	if err != nil {
 		return err
 	}
 
-	authFile, cleaner, err := shared_podman.PodmanLogin(hostData, flags.Image.Registry, flags.Installation.SCC)
+	authFile, cleaner, err := podman_utils.PodmanLogin(hostData, flags.Image.Registry, flags.Installation.SCC)
 	if err != nil {
 		return err
 	}
@@ -34,7 +40,7 @@ func upgradePodman(_ *types.GlobalFlags, flags *podmanUpgradeFlags, cmd *cobra.C
 		return errors.New(L("install podman before running this command"))
 	}
 
-	return podman.Upgrade(
+	return podman.Migrate(
 		systemd, authFile,
 		flags.Installation.DB,
 		flags.Installation.ReportDB,
@@ -45,6 +51,10 @@ func upgradePodman(_ *types.GlobalFlags, flags *podmanUpgradeFlags, cmd *cobra.C
 		flags.HubXmlrpc,
 		flags.Saline,
 		flags.Pgsql,
-		flags.Installation.TZ,
+		flags.Migration.Prepare,
+		flags.Migration.User,
+		flags.Mirror,
+		flags.Podman,
+		args,
 	)
 }
